+++ lib/export/coldp/files/name_relation.rb
@@ -105,6 +105,7 @@ module Export::Coldp::Files::NameRelation
         modified
         modifiedBy
         remarks
+        tw_taxon_name_relationship_id
       }
 
       rels = taxon_name_relationships(otu)
@@ -121,6 +122,7 @@ module Export::Coldp::Files::NameRelation
             Export::Coldp.modified(tnr[:updated_at]),                        # modified
             Export::Coldp.modified_by(tnr[:updated_by_id], project_members), # modified_by
             nil,                                                             # remarks
+            tnr.id                                                           # tw_taxon_name_relationship_id
           ]
         end
       end